feat: 모킹 api 로직 분리
