+++ packages/vanilla/src/main-server.js
@@ -5,104 +5,7 @@ import { PRODUCT_ACTIONS } from "./stores/actionTypes.js";
 import { uiStore } from "./stores/uiStore.js";
 import { route } from "./router/serverRouter.js";
 import { router } from "./router/router.js";
-import items from "./mocks/items.json";
-
-function getUniqueCategories() {
-  const categories = {};
-  items.forEach((item) => {
-    const cat1 = item.category1;
-    const cat2 = item.category2;
-    if (!categories[cat1]) categories[cat1] = {};
-    if (cat2 && !categories[cat1][cat2]) categories[cat1][cat2] = {};
-  });
-  return categories;
-}
-
-function filterProducts(products, query) {
-  let filtered = [...products];
-
-  if (query.search) {
-    const searchTerm = query.search.toLowerCase();
-    filtered = filtered.filter(
-      (item) => item.title.toLowerCase().includes(searchTerm) || item.brand.toLowerCase().includes(searchTerm),
-    );
-  }
-
-  if (query.category1) {
-    filtered = filtered.filter((item) => item.category1 === query.category1);
-  }
-  if (query.category2) {
-    filtered = filtered.filter((item) => item.category2 === query.category2);
-  }
-
-  if (query.sort) {
-    switch (query.sort) {
-      case "price_asc":
-        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
-        break;
-      case "price_desc":
-        filtered.sort((a, b) => parseInt(b.lprice) - parseInt(a.lprice));
-        break;
-      case "name_asc":
-        filtered.sort((a, b) => a.title.localeCompare(b.title, "ko"));
-        break;
-      case "name_desc":
-        filtered.sort((a, b) => b.title.localeCompare(a.title, "ko"));
-        break;
-      default:
-        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
-    }
-  }
-
-  return filtered;
-}
-
-async function mockGetProducts(query = {}) {
-  const normalized = {
-    page: Number(query?.page ?? query?.current ?? 1),
-    limit: Number(query?.limit ?? 20),
-    search: query?.search ?? "",
-    category1: query?.category1 ?? "",
-    category2: query?.category2 ?? "",
-    sort: query?.sort ?? "price_asc",
-  };
-
-  const filteredProducts = filterProducts(items, normalized);
-
-  const uptoEnd = Math.min(filteredProducts.length, normalized.page * normalized.limit);
-  const productsForResponse = filteredProducts.slice(0, uptoEnd);
-
-  return {
-    products: productsForResponse,
-    pagination: {
-      page: normalized.page,
-      limit: normalized.limit,
-      total: filteredProducts.length,
-      totalPages: Math.ceil(filteredProducts.length / normalized.limit),
-      hasNext: uptoEnd < filteredProducts.length,
-      hasPrev: normalized.page > 1,
-    },
-  };
-}
-
-async function mockGetProduct(id) {
-  const product = items.find((item) => item.productId === id);
-  if (!product) return null;
-
-  return {
-    ...product,
-    description: `${product.title}에 대한 상세 설명입니다.`,
-    rating: Math.floor(Math.random() * 2) + 4,
-    reviewCount: Math.floor(Math.random() * 1000) + 50,
-    stock: Math.floor(Math.random() * 100) + 10,
-    images: [product.image, product.image.replace(".jpg", "_2.jpg"), product.image.replace(".jpg", "_3.jpg")],
-  };
-}
-
-async function mockGetCategories() {
-  const categories = getUniqueCategories();
-  return categories;
-}
+import { mockGetProducts, mockGetProduct, mockGetCategories } from "./mocks/mockApi.js";
 
 route.add("/", async ({ query }) => {
   const productsData = await mockGetProducts(query || {});
